+++ src/java/components/safetylogic/hu.bme.mit.inf.modes3.components.safetylogic.componentlevel.sc/src/main/java/org/yakindu/scr/turnout/ITurnoutStatemachine.java
@@ -49,13 +49,13 @@ public interface ITurnoutStatemachine extends IStatemachine {
 
 		public boolean isRaisedReleaseBottom();
 
-		public long getId();
+		public int getId();
 
-		public void setId(long value);
+		public void setId(int value);
 
-		public long getCount();
+		public int getCount();
 
-		public void setCount(long value);
+		public void setCount(int value);
 
 		public List<SCITurnoutListener> getListeners();
